Foram feitas alterações no Mapper de 4 entidades
- Por causa de alterações no DTO, alterou-se o Mapper de Comment, Contact, Project e Review para ficar de acordo com o novo construtor;
-Alterou-se o tipo das enum no DTO de Project para int.

+++ src/main/java/br/senac/bluhelp/dto/project/ProjectDTO.java
@@ -1,4 +1,4 @@
 package br.senac.bluhelp.dto.project;
 
-public record ProjectDTO(Long id, Long creator, String title, String objective, Long address, String description, Long progress, Long category, byte[] photo) {
+public record ProjectDTO(Long id, Long creator, String title, String objective, Long address, String description, int progress, int category, byte[] photo) {
 }
